+++ clease/montecarlo/__init__.py
@@ -1,8 +1,10 @@
 # pylint: disable=undefined-variable
+from .mc_evaluator import *
 from .barrier_models import *
 from .kmc_events import *
 from .kinetic_monte_carlo import *
 from .trial_move_generator import *
+from . import base
 from .bias_potential import BiasPotential
 from .montecarlo import Montecarlo
 from .sgc_montecarlo import SGCMonteCarlo
@@ -12,6 +14,6 @@ from .random_bias_with_memory import RandomBiasWithMemory
 from .metadynamics_sampler import MetaDynamicsSampler
 
 ADDITIONAL = ('Montecarlo', 'SGCMonteCarlo', 'MetaDynamicsSampler', 'BinnedBiasPotential',
-              'BiasPotential', 'GaussianKernelBiasPotential', 'RandomBiasWithMemory')
+              'BiasPotential', 'GaussianKernelBiasPotential', 'RandomBiasWithMemory', 'base')
 __all__ = ADDITIONAL + (barrier_models.__all__ + kinetic_monte_carlo.__all__ + kmc_events.__all__ +
-                        trial_move_generator.__all__)
+                        trial_move_generator.__all__ + mc_evaluator.__all__)