+++ clease/montecarlo/montecarlo.py
@@ -1,40 +1,56 @@
-# -*- coding: utf-8 -*-
 """Monte Carlo method for ase."""
+from typing import Tuple, Dict, Union
 import sys
 import datetime
 import time
 import logging
 from collections import Counter
-from typing import Sequence
 import numpy as np
 from ase import Atoms
 from ase.units import kB
-from clease.datastructures import SystemChange
+import clease
+from clease.datastructures import SystemChange, SystemChanges
+from clease.montecarlo.mc_evaluator import MCEvaluator
 from clease.montecarlo.exponential_filter import ExponentialFilter
 from clease.montecarlo.averager import Averager
 from clease.montecarlo.bias_potential import BiasPotential
 from clease.montecarlo.observers import MCObserver
 from clease.montecarlo.trial_move_generator import TrialMoveGenerator, RandomSwap
+from .base import BaseMC
 
 logger = logging.getLogger(__name__)
 
 
 # pylint: disable=too-many-instance-attributes
-class Montecarlo:
+class Montecarlo(BaseMC):
     """Class for running Monte Carlo at a fixed composition (canonical).
 
-    :param atoms: ASE Atoms object (with CE calculator attached)
-    :param temp: Temperature of Monte Carlo simulation in Kelvin
-    :param generator: Generator that produces trial moves
+    Args:
+        system (Union[ase.Atoms, MCEvaluator]): Either an ASE Atoms object
+            with an attached calculator, or a pre-initialized
+            :class:`~clease.montecarlo.mc_evaluator.MCEvaluator`
+            object.
+        temp (float): Temperature of Monte Carlo simulation in Kelvin
+        generator (TrialMoveGenerator, optional): A
+            :class:`~clease.montecarlo.trial_move_generator.TrialMoveGenerator`
+            object that produces trial moves. Defaults to None.
+        rng (numpy.random.Generator, optional): NumPy RNG generator object to use
+            for the random number generation. Defaults to None.
     """
+    NAME = "MonteCarlo"
 
-    def __init__(self, atoms: Atoms, temp: float, generator: TrialMoveGenerator = None):
-        self.name = "MonteCarlo"
-        self.atoms = atoms
+    def __init__(self,
+                 system: Union[Atoms, MCEvaluator],
+                 temp: float,
+                 generator: TrialMoveGenerator = None,
+                 rng: np.random.Generator = None):
+
+        super().__init__(system, rng=rng)
         self.T = temp
 
         if generator is None:
-            self.generator = RandomSwap(atoms)
+            # If rng is None, RandomSwap will create its own rng object.
+            self.generator = RandomSwap(self.atoms, rng=rng)
         else:
             self.generator = generator
 
@@ -46,17 +62,15 @@ class Montecarlo:
         self.current_step = 0
         self.num_accepted = 0
         self.status_every_sec = 30
-        E0 = self.atoms.calc.calculate(None, None, None)
-        self.atoms.calc.clear_history()
-        self.current_energy = E0
-        self.new_energy = self.current_energy
 
-        self.trial_move = []  # Last trial move performed
-        self.mean_energy = Averager(ref_value=self.current_energy)
-        self.energy_squared = Averager(ref_value=self.current_energy)
+        self.trial_move = []
+        # We cannot cause an energy calculation trigger in init,
+        # so we defer these quantities until needed.
+        self.current_energy = None
+        self.new_energy = None
+        self.mean_energy = None
+        self.energy_squared = None
 
-        # Some member variables used to update the atom tracker, only relevant
-        # for canonical MC
         self.quit = False
 
         self.filter = ExponentialFilter(min_time=0.2 * len(self.atoms),
@@ -64,8 +78,22 @@ class Montecarlo:
                                         n_subfilters=10)
 
     def update_current_energy(self):
-        """Enforce a new energy evaluation."""
-        self.current_energy = self.atoms.get_potential_energy()
+        self.current_energy = self.evaluator.get_energy()
+        logger.debug('Updating current energy to %s', self.current_energy)
+        self.evaluator.reset()
+
+    def _initialize_energies(self) -> None:
+        """Initialize the energy quantities, causes a trigger of an energy evaluation."""
+
+        self.update_current_energy()
+        self.new_energy = self.current_energy
+
+        # Create new averager objects, only the first time
+        # (since they are initialized as None in __init__)
+        if self.mean_energy is None:
+            self.mean_energy = Averager(ref_value=self.current_energy)
+        if self.energy_squared is None:
+            self.energy_squared = Averager(ref_value=self.current_energy)
 
     def reset(self):
         """Reset all member variables to their original values."""
@@ -73,11 +101,14 @@ class Montecarlo:
         for _, obs in self.observers:
             obs.reset()
 
+        self.evaluator.reset()
         self.filter.reset()
         self._reset_internal_counters()
 
-        self.mean_energy.clear()
-        self.energy_squared.clear()
+        if self.mean_energy is not None:
+            self.mean_energy.clear()
+        if self.energy_squared is not None:
+            self.energy_squared.clear()
 
     def _reset_internal_counters(self):
         """Reset the step counters which are used internally"""
@@ -113,15 +144,16 @@ class Montecarlo:
 
         self.observers.append((interval, obs))
 
-    def _initialize_run(self):
+    def initialize_run(self):
+        """Prepare MC object for a new run."""
+        logger.debug('Initializing run')
         self.generator.initialize(self.atoms)
 
-        self.update_current_energy()
+        # Initialize/update relevant energy quantities
+        self._initialize_energies()
+
         # Reset the internal step counters
         self._reset_internal_counters()
-        # Atoms object should have attached calculator
-        # Add check that this is show
-        self._mc_step()
 
     def run(self, steps: int = 100):
         """Run Monte Carlo simulation.
@@ -131,19 +163,23 @@ class Montecarlo:
         steps: int
             Number of steps in the MC simulation
         """
+
+        logger.info('Starting MC run with %d steps.', steps)
         # Check the number of different elements are correct to avoid
         # infinite loops
-        self._initialize_run()
+        self.initialize_run()
 
         start = time.time()
         prev = self.current_step
+        info_enabled = logger.isEnabledFor(logging.INFO)  # Do we emit INFO logs?
         while self.current_step < steps:
             E, _ = self._mc_step()
 
             self.mean_energy += E
             self.energy_squared += E**2
 
-            if time.time() - start > self.status_every_sec:
+            # We only want to do this calculation if logging is enabled for INFO.
+            if info_enabled and time.time() - start > self.status_every_sec:
                 ms_per_step = 1000.0 * self.status_every_sec / (self.current_step - prev)
                 accept_rate = self.num_accepted / self.current_step
                 logger.info("%d of %d steps. %.2f ms per step. Acceptance rate: %.2f",
@@ -152,7 +188,7 @@ class Montecarlo:
                 start = time.time()
 
             if self.quit:
-                logger.debug('Quit signal detected. Breaking.')
+                logger.warning('Quit signal detected. Breaking the MC loop.')
                 break
 
         logger.info("Reached maximum number of steps (%d mc steps)", steps)
@@ -162,10 +198,12 @@ class Montecarlo:
         """Return dict with meta info."""
         ts = time.time()
         st = datetime.datetime.fromtimestamp(ts).strftime('%Y-%m-%d %H:%M:%S')
+        clease_version = clease.version.__version__
         v_info = sys.version_info
         meta_info = {
             "timestamp": st,
-            "python_version": f"{v_info.major}.{v_info.minor}.{v_info.micro}"
+            "python_version": f"{v_info.major}.{v_info.minor}.{v_info.micro}",
+            "clease_version": clease_version,
         }
         return meta_info
 
@@ -175,8 +213,7 @@ class Montecarlo:
         mean_energy = self.mean_energy.mean
         quantities["energy"] = mean_energy
         mean_sq = self.energy_squared.mean
-        quantities["heat_capacity"] = \
-            (mean_sq - mean_energy**2) / (kB * self.T**2)
+        quantities["heat_capacity"] = (mean_sq - mean_energy**2) / (kB * self.T**2)
         quantities["energy_var"] = mean_sq - mean_energy**2
         quantities["temperature"] = self.T
         at_count = self.count_atoms()
@@ -193,7 +230,7 @@ class Montecarlo:
             quantities.update(obs[1].get_averages())
         return quantities
 
-    def _calculate_step(self, system_changes: Sequence[SystemChange]):
+    def _calculate_step(self, system_changes: SystemChanges):
         """Calculate energies given a step, and decide if we accept the step.
 
         Returns boolean if system changes are accepted.
@@ -203,24 +240,28 @@ class Montecarlo:
         system_changes: list
             List with system changes
         """
-        # NOTE: Calculate updates the system
-        calc = self.atoms.calc
+        logger.debug('Applying changes to the system.')
+        self.evaluator.apply_system_changes(system_changes)
+        # Evaluate the energy quantity after applying the changes to the system.
+        self.new_energy = self.evaluator.get_energy(applied_changes=system_changes)
 
-        with calc.with_system_changes(system_changes) as keeper:
-            self.new_energy = calc.get_energy()
+        # NOTE: As this is called after calculate, the changes has
+        # already been introduced to the system
+        for bias in self.bias_potentials:
+            self.new_energy += bias(system_changes)
+        logger.debug('Current energy: %.3f eV, new energy: %.3f eV', self.current_energy,
+                     self.new_energy)
+        accept = self._do_accept(self.current_energy, self.new_energy)
+        logger.debug('Change accepted? %s', accept)
 
-            # NOTE: As this is called after calculate, the changes has
-            # already been introduced to the system
-            for bias in self.bias_potentials:
-                self.new_energy += bias(system_changes)
+        if accept:
+            # Changes accepted, finalize evaluator.
+            self.evaluator.keep_system_changes(system_changes)
+        else:
+            # Undo changes
+            logger.debug('Change rejected, undoing system changes.')
+            self.evaluator.undo_system_changes(system_changes)
 
-            logger.debug('Current energy: %.3f eV, new energy: %.3f eV', self.current_energy,
-                         self.new_energy)
-            accept = self._do_accept(self.current_energy, self.new_energy)
-            logger.debug('Change accepted? %s', accept)
-
-            # Decide if we keep changes, or rollback
-            keeper.keep_changes = accept
         return accept
 
     def _do_accept(self, current_energy: float, new_energy: float) -> bool:
@@ -239,10 +280,11 @@ class Montecarlo:
         kT = self.T * kB
         energy_diff = new_energy - current_energy
         probability = np.exp(-energy_diff / kT)
-        logger.debug('Calculated probability: %.3f', probability)
-        return np.random.rand() <= probability
+        logger.debug('Energy difference: %.3e. Calculated probability: %.3f', energy_diff,
+                     probability)
+        return self.rng.random() <= probability
 
-    def _move_accepted(self, system_changes: Sequence[SystemChange]) -> Sequence[SystemChange]:
+    def _move_accepted(self, system_changes: SystemChanges) -> SystemChanges:
         logger.debug('Move accepted, updating things')
         self.num_accepted += 1
         self.generator.on_move_accepted(system_changes)
@@ -250,7 +292,7 @@ class Montecarlo:
         return system_changes
 
     # pylint: disable=no-self-use
-    def _move_rejected(self, system_changes: Sequence[SystemChange]) -> Sequence[SystemChange]:
+    def _move_rejected(self, system_changes: SystemChanges) -> SystemChanges:
         logger.debug('Move rejected, undoing system changes: %s', system_changes)
         self.generator.on_move_rejected(system_changes)
 
@@ -264,11 +306,11 @@ class Montecarlo:
         logger.debug('Reversed system changes: %s', system_changes)
         return system_changes
 
-    def count_atoms(self):
+    def count_atoms(self) -> Dict[str, int]:
         """Count the number of each element."""
         return dict(Counter(self.atoms.symbols))
 
-    def _mc_step(self):
+    def _mc_step(self) -> Tuple[float, bool]:
         """Make one Monte Carlo step by swithing two atoms."""
         self.current_step += 1
 
@@ -286,7 +328,7 @@ class Montecarlo:
         self.filter.add(self.current_energy)
         return self.current_energy, move_accepted
 
-    def execute_observers(self, system_changes: Sequence[SystemChange]):
+    def execute_observers(self, system_changes: SystemChanges):
         for interval, obs in self.observers:
             if self.current_step % interval == 0:
                 logger.debug('Executing observer %s at step %d with interval %d.', obs.name,